+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Adds custom `useCarouselStyles` hook for the `Tabs` mobile functionality.
 - Adds custom NYPL-theme radii values to use within Chakra.
+- Adds the `showHelperInvalidText` prop to form components to control displaying or hiding the content in the `HelperErrorText` component.
+- Composes the DS `Notification` component with Chakra components.
+- Composes the DS `HelperErrorText` component with Chakra.
 - Composes the DS `List` component with Chakra.
 - Composes the DS `Pagination` component with Chakra.